% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_documents.r
\name{browse_gtd}
\alias{browse_gtd}
\title{Opens a GTD event in the browser}
\usage{
browse_gtd(id)
}
\arguments{
\item{id}{A GTD event id}
}
\value{
Opens browser window
}
\description{
Opens a GTD event in the browser
}
