% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inspect_documents.r
\name{browse_guardian}
\alias{browse_guardian}
\title{Opens a Guardian article in the browser}
\usage{
browse_guardian(id)
}
\arguments{
\item{id}{A Guardian article id}
}
\value{
Opens browser window
}
\description{
Opens a Guardian article in the browser
}
