% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/est_validity.r
\name{calc_chi2}
\alias{calc_chi2}
\title{Vectorized computation of chi^2 statistic for a 2x2 crosstab containing the values
[a, b]
[c, d]}
\usage{
calc_chi2(a, b, c, d, correct = T)
}
\arguments{
\item{a}{topleft value of the table}

\item{b}{topright value}

\item{c}{bottomleft value}

\item{d}{bottomright value}

\item{correct}{if TRUE, use yates correction. Can be a vector of length a (i.e. the number of tables)}
}
\description{
Vectorized computation of chi^2 statistic for a 2x2 crosstab containing the values
[a, b]
[c, d]
}
