% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corenlp_parse.r, R/rsyntax_recipe.r
\name{corenlp_clauses}
\alias{corenlp_clauses}
\title{Function that wraps tqueries for extracting clauses}
\usage{
corenlp_clauses(tokens, verbs = NULL, exclude_verbs = NULL)

corenlp_clauses(tokens, verbs = NULL, exclude_verbs = NULL)
}
\arguments{
\item{tokens}{A tokenindex}

\item{verbs}{A character vector with verbs (as lemma) that indicate clauses}

\item{exclude_verbs}{A character vector with verbs to explicitly exclude as clause}

\item{tokens}{A tokenindex}

\item{verbs}{A character vector with verbs (as lemma) that indicate clauses}

\item{exclude_verbs}{A character vector with verbs to explicitly exclude as clause}
}
\value{
A tokenindex

A tokenindex
}
\description{
Function that wraps tqueries for extracting clauses

Function that wraps tqueries for extracting clauses
}
