% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corenlp_parse.r, R/rsyntax_recipe.r
\name{corenlp_quotes}
\alias{corenlp_quotes}
\title{Function that wraps tqueries for extracting quotes}
\usage{
corenlp_quotes(tokens, verbs, exclude_verbs = NULL)

corenlp_quotes(tokens, verbs, exclude_verbs = NULL)
}
\arguments{
\item{tokens}{A tokenindex}

\item{verbs}{A character vector with verbs (as lemma) that indicate speech acts}

\item{exclude_verbs}{A character vector with verbs to explicitly exclude as speech acts}

\item{tokens}{A tokenindex}

\item{verbs}{A character vector with verbs (as lemma) that indicate speech acts}

\item{exclude_verbs}{A character vector with verbs to explicitly exclude as speech acts}
}
\value{
A tokenindex

A tokenindex
}
\description{
Function that wraps tqueries for extracting quotes

Function that wraps tqueries for extracting quotes
}
