% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corenlp_parse.r, R/rsyntax_recipe.r
\name{corenlp_simplify}
\alias{corenlp_simplify}
\title{Simplify the dependency tree by unpacking conjunctions and relative clauses}
\usage{
corenlp_simplify(tokens)

corenlp_simplify(tokens)
}
\arguments{
\item{tokens}{A tokenindex (as_tokenindex)}

\item{tokens}{A tokenindex (as_tokenindex)}
}
\value{
A tokenindex

A tokenindex
}
\description{
Simplify the dependency tree by unpacking conjunctions and relative clauses

Simplify the dependency tree by unpacking conjunctions and relative clauses
}
