% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoparsing.r
\name{geo_tags}
\alias{geo_tags}
\title{Create geo tags}
\usage{
geo_tags(txt, min_freq = NA, top_n = NA, batchsize = 10000,
  with_model = F)
}
\arguments{
\item{txt}{A character vector with texts}

\item{min_freq}{Minimum frequency of matches in the geo dictionary}

\item{top_n}{If used, return only the top n tags}

\item{batchsize}{Calculation is performed in batches}
}
\value{
A character vector of length txt with geo tags, concatenated with a komma delimiter
}
\description{
Uses the newsmap package to create geo tags. Currenly only simple dictionary matching is used.
}
\examples{
geo_tags('Paris Germany Berlin Leiden')
}
