% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{gtd_comparison_results}
\alias{gtd_comparison_results}
\title{GTD matching results}
\format{An RNewsflow_edgelist object, which is a list with an edgelist (i.e. GTD -> news article pairs) and document meta data.}
\usage{
data(terrorism_terms)
}
\description{
This is the result of the compare_documents function as used in the vignette.
}
\examples{
gtd_comparison_results
}
\keyword{datasets}
