% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gtd_match_analysis.r
\name{gtd_event_coverage}
\alias{gtd_event_coverage}
\title{Extract GTD events with news article annotations}
\usage{
gtd_event_coverage(g, weight_thres, use_geomatch)
}
\arguments{
\item{use_geomatch}{}
}
\description{
Extract GTD events with news article annotations
}
