% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoparsing.r
\name{match_geo}
\alias{match_geo}
\title{Match geo code}
\usage{
match_geo(g, from_var = "geo", to_var = "geo", sep = ",")
}
\arguments{
\item{g}{An edgelist, as created with newsflow.compare}

\item{from_var}{The name of the geo code column in from_dtm}

\item{to_var}{The name of the geo code column in to_dtm}
}
\value{
A logical vector
}
\description{
If the DTMs used in the compare_documents function contained docvars with geo codes,
it can be used to match documents based on their geo code. This can be a single code (e.g., "en")
or multiple codes concatenated with a comma (e.g., "en,de") or another seperator defined in the 'sep' argument.
A match is TRUE if at least one of the codes in the from document is the same as at least one of the codes in the to document.
}
\details{
The function returns a logical vector indicating if each match in the edge list has a geo match.
}
