% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_data.r
\name{prepare_gtd}
\alias{prepare_gtd}
\title{Prepare the GTD data}
\usage{
prepare_gtd(d, fromdate = NULL, todate = NULL,
  cols = "summary|^city|country_txt|attacktype[0-9]\\\\_txt|targtype[0-9]\\\\_txt|targsubtype[0-9]\\\\_txt|natlty[0-9]\\\\_txt|gname[0-9]|gsubname[0-9]|weaptype[0-9]\\\\_txt|weapdetail|victims",
  with_geo = T)
}
\arguments{
\item{d}{A data.frame with the GTD data. You can download the GTD data for free
by filling out \href{https://www.start.umd.edu/gtd/contact/}{this form}.
You can read the .xlsx file with \code{\link[openxlsx]{read.xlsx}}.}

\item{fromdate}{A Date (or character value that can be coerced to Date)}

\item{todate}{A Date (or character value that can be coerced to Date)}
}
\value{
A quanteda style DTM
}
\description{
Creates a document term matrix in the \code{\link[quanteda]{dfm}} class.
}
\examples{
\donttest{
gtd_data = openxlsx::read.xlsx('GTD FILE AS DOWNLOADED FROM GTD WEBSITE')
gtd_dtm = prepare_gtd(gtd_data, fromdate='2010-01-01')
}
}
