% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corenlp_parse.r
\name{sayverbs}
\alias{sayverbs}
\title{Get a vector of english verbs that indicate speech}
\usage{
sayverbs(add_verbs = NULL, rm_verbs = NULL)
}
\arguments{
\item{add_verbs}{add verbs}

\item{rm_verbs}{remove verbs}
}
\value{
A character vector
}
\description{
Get a vector of english verbs that indicate speech
}
\examples{
sayverbs()
}
