% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_tquery_annotations.r
\name{spacy_actions}
\alias{spacy_actions}
\title{Function that wraps tqueries for extracting actions}
\usage{
spacy_actions(tokens)
}
\arguments{
\item{tokens}{A tokenindex}
}
\value{
A tokenindex
}
\description{
Actions
}
