% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_tquery_annotations.r
\name{spacy_links}
\alias{spacy_links}
\title{Function that wraps tqueries for extracting links}
\usage{
spacy_links(tokens)
}
\arguments{
\item{tokens}{A tokenindex}
}
\value{
A tokenindex
}
\description{
Links of subjects to descriptions. This can be with the linking verb "be" ("Bob is sick"),
but also with appositional modifiers (Bob, the director, argued that...") or ajectival clauses ("A fish called Wanda")
}
