% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_tquery_annotations.r
\name{spacy_poss}
\alias{spacy_poss}
\title{Function that wraps tqueries for extracting possive relations}
\usage{
spacy_poss(tokens)
}
\arguments{
\item{tokens}{A tokenindex}
}
\value{
A tokenindex
}
\description{
Links of subjects to possesives. This can be with the linking verb "have" ("Bob has cookies"),
but also with the poss relation.
}
