% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_tquery_annotations.r
\name{spacy_quotes}
\alias{spacy_quotes}
\title{Function that wraps tqueries for extracting quotes}
\usage{
spacy_quotes(tokens)
}
\arguments{
\item{tokens}{A tokenindex}
}
\value{
A tokenindex
}
\description{
Function that wraps tqueries for extracting quotes
}
