% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_tree_transformations.r
\name{spacy_simplify}
\alias{spacy_simplify}
\title{Simplify the dependency tree}
\usage{
spacy_simplify(tokens)
}
\arguments{
\item{tokens}{A tokenindex (as_tokenindex)}
}
\value{
A tokenindex
}
\description{
Simplify the dependency tree
}
