% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacy_tquery_annotations.r
\name{spacy_verb_clauses}
\alias{spacy_verb_clauses}
\title{Function that wraps tqueries for extracting verb clauses}
\usage{
spacy_verb_clauses(tokens, column, verbs = NULL, exclude_verbs = NULL,
  ...)
}
\arguments{
\item{tokens}{A tokenindex}

\item{column}{The name of the column with the annotations}

\item{verbs}{Optionally, a character vector to specify specific verb lemma}

\item{exclude_verbs}{Optionally, a character vector to specify verb lemma to exclude}

\item{...}{additional tqueries to annotate}
}
\value{
A tokenindex
}
\description{
Function that wraps tqueries for extracting verb clauses
}
