% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_par.r
\name{suspect_comm_paragraphs}
\alias{suspect_comm_paragraphs}
\title{Extract paragraphs that mention islam and terrorism from a text}
\usage{
suspect_comm_paragraphs(text, islam_regex = "islam|muslim",
  terror_regex = "terror", ignore.case = T, ...)
}
\arguments{
\item{islam_regex}{A regular expression for findings texts that mention islam}

\item{terror_regex}{A regular expression for finding texts that mention terrorism}

\item{ignore.case}{Passed to grepl}

\item{...}{Additional arguments passed to grepl}

\item{corp}{A quanteda corpus}
}
\value{
A quanteda corpus
}
\description{
Extract paragraphs that mention islam and terrorism from a text
}
\examples{
corp = quanteda::corpus(data.frame(stringsAsFactors = F,
id = c(1,2),
date = c('2010-01-01','2010-02-01'),
text = c('terror islam.\\n\\ntest.\\n\\nterrorism muslim', 'test.\\n\\nterrorism islam')
))

corp2 = suspect_comm_paragraphs(corp)
corp2$documents
}
