% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoparsing.r
\name{match_geo}
\alias{match_geo}
\title{Match geo code}
\usage{
match_geo(g, from_dtm, to_dtm, from_var = "geo", to_var = "geo")
}
\arguments{
\item{g}{An edgelist, as created with newsflow.compare}

\item{from_dtm}{The DTM (quanteda dfm) for the articles in the `from` column in g}

\item{to_dtm}{The DTM (quanteda dfm) for the articles in the `to` column in g}

\item{from_var}{The name of the geo code column in from_dtm}

\item{to_var}{The name of the geo code column in to_dtm}
}
\value{
A logical vector
}
\description{
The DTMs created with \code{\link{prepare_gtd}} and \code{\link{prepare_news}} have a "geo" document variable.
The match_geo function uses this information to match documents based on their geo code. This can be a single code (e.g., "en")
or multiple codes concatenated with a comma (e.g., "en,de").
}
\details{
The function returns a logical vector indicating if each match in the edge list has a geo match.
}
